refactor: データ構造を別ファイルに移動
