+++ akanec/src/data/token.rs
@@ -1,0 +1,11 @@
+
+#[derive(Clone, Debug, PartialEq)]
+pub enum Token {
+    Eof,
+    Ident(String),
+    Num(String),
+    OpCode(String),
+    Equal,
+    LParen,
+    RParen,
+}
