+++ akanec/src/main.rs
@@ -1,3 +1,4 @@
+mod data;
 mod lexer;
 mod parser;
 mod llvm;
